static casted the string_view.size() to 'int' for correcteness.

+++ src/AzslcSemanticOrchestrator.cpp
@@ -570,7 +570,7 @@ namespace AZ::ShaderCompiler
             if (!varInfo.m_typeInfoExt.IsClassFound())
             {
                 const string message = FormatString("Unknown type '%s' for shader option '%.*s'",
-                    varInfo.m_typeInfoExt.GetDisplayName().c_str(), uqNameView.size(), uqNameView.data());
+                    varInfo.m_typeInfoExt.GetDisplayName().c_str(), static_cast<int>(uqNameView.size()), uqNameView.data());
                 ThrowAzslcOrchestratorException(ORCHESTRATOR_UNKNOWN_OPTION_TYPE, ctx->start, message);
             }
             // we'll set that here, an option is better flagged as static const for simplicity during emission
